Collapse the triplicated baseline analysis model into shared code (Lite + Darling)
The anomaly brain's baseline MODEL was carried verbatim in three places
(Lite BaselineProvider/AnomalyDetector, Darling Pg*, Dashboard SqlServer*) --
a drift risk AnomalyGate itself flagged. This is the last Tier-0 parity
forcing-function item.

Extract to shared PerformanceMonitor.Analysis.Baselines, consumed by BOTH
active apps:
- BaselineBucket (+ EffectiveStdDev/IsTrustworthy + tier-gate floors)
- BaselineTier
- BaselineMath: tier selection w/ hysteresis (SelectBucket), hour-only/flat
  pooling, pooled variance, dispersion floor, window/collapse/restore thresholds
- AnomalyThresholds: the ~20 detector tuning constants
- MetricNames: baseline cache keys

Each provider keeps only its dialect-specific baseline SQL + parameter binding.

Verified byte-identical between Lite and Darling before extracting; not one
threshold, constant, or line of logic changed. The .Baselines sub-namespace
avoids a CS0104 collision with the deprecated Dashboard's own same-named
BaselineBucket/BaselineTier -- the Dashboard is untouched and compiles unchanged.

Adds SharedBaselineModelPinTests on each side asserting each app binds the
model to the shared assembly and holds no per-app copy (can't re-fork).

Full Lite + Darling baseline/anomaly suites pass unchanged.
