+++ Darling/Darling.Tests/DarlingAnomalyBaselineTests.cs
@@ -13,6 +13,7 @@ using System.Reflection;
 using System.Threading.Tasks;
 using Npgsql;
 using PerformanceMonitor.Analysis;
+using PerformanceMonitor.Analysis.Baselines;
 using PerformanceMonitor.Darling.Analysis;
 using PerformanceMonitor.Darling.Storage;
 using Xunit;
