+++ Lite/Analysis/BaselineProvider.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using DuckDB.NET.Data;
-using PerformanceMonitor.Analysis;
+using PerformanceMonitor.Analysis.Baselines;
 using PerformanceMonitorLite.Database;
 using PerformanceMonitorLite.Services;
 
@@ -26,15 +26,6 @@ namespace PerformanceMonitorLite.Analysis;
 public class BaselineProvider
 {
     private readonly DuckDbInitializer _duckDb;
-
-    /// <summary>Rolling window for baseline computation.</summary>
-    private const int BaselineWindowDays = 30;
-
-    /// <summary>Collapse to hour-only when full bucket has fewer than this many samples.</summary>
-    private const int CollapseThreshold = 10;
-
-    /// <summary>Restore to full bucket when sample count reaches this level (hysteresis).</summary>
-    private const int RestoreThreshold = 15;
 
     /// <summary>Cache TTL — baselines are recomputed after this interval.</summary>
     public static TimeSpan CacheTtl { get; set; } = TimeSpan.FromHours(1);
@@ -60,39 +51,7 @@ public class BaselineProvider
         if (baselines == null || baselines.Count == 0)
             return BaselineBucket.Empty;
 
-        // Try full bucket (hour + day-of-week)
-        var fullKey = (hourOfDay, dayOfWeek);
-        if (baselines.TryGetValue(fullKey, out var fullBucket) && fullBucket.SampleCount >= RestoreThreshold)
-            return fullBucket;
-
-        // If full bucket exists but below restore threshold, check if it's above collapse threshold
-        // (hysteresis: don't collapse if we're between 10-14 samples and were previously using full)
-        if (fullBucket != null && fullBucket.SampleCount >= CollapseThreshold)
-            return fullBucket;
-
-        // Collapse to hour-only: aggregate all days for this hour
-        var hourBuckets = baselines
-            .Where(kvp => kvp.Key.HourOfDay == hourOfDay)
-            .Select(kvp => kvp.Value)
-            .ToList();
-
-        if (hourBuckets.Count > 0)
-        {
-            var collapsed = CollapseToHourOnly(hourBuckets);
-            if (collapsed.SampleCount >= CollapseThreshold)
-                return collapsed;
-        }
-
-        // Collapse to flat: aggregate everything
-        var allBuckets = baselines.Values.ToList();
-        if (allBuckets.Count > 0)
-        {
-            var flat = CollapseToFlat(allBuckets);
-            if (flat.SampleCount >= 3) // Minimum viable baseline
-                return flat;
-        }
-
-        return BaselineBucket.Empty;
+        return BaselineMath.SelectBucket(baselines, hourOfDay, dayOfWeek);
     }
 
     /// <summary>Forces cache eviction for a server — used during testing.</summary>
@@ -137,8 +96,8 @@ public class BaselineProvider
         var query = GetBaselineQuery(metricName);
         if (query == null) return null;
 
-        var absStdDevFloor = AbsStdDevFloorFor(metricName);
-        var windowStart = analysisTime.AddDays(-BaselineWindowDays);
+        var absStdDevFloor = BaselineMath.AbsStdDevFloorFor(metricName);
+        var windowStart = analysisTime.AddDays(-BaselineMath.BaselineWindowDays);
 
         try
         {
@@ -400,216 +359,10 @@ GROUP BY hour_of_day, day_of_week",
         };
     }
 
-    /// <summary>
-    /// Bounded-metric absolute dispersion floor (see BaselineBucket.AbsStdDevFloor). Server-relative
-    /// metrics have no universal floor and return 0. Tunable — calibrate on the SQL2025/HammerDB box.
-    /// </summary>
-    private static double AbsStdDevFloorFor(string metricName) => metricName switch
-    {
-        MetricNames.Cpu => 5.0,        // CPU utilization %
-        MetricNames.Memory => 4.0,     // memory pressure %
-        MetricNames.IoLatency => 2.5,  // I/O latency ms
-        _ => 0.0,                       // batch/query-duration/sessions/waits/blocking/deadlock — server-relative
-    };
-
-    /// <summary>
-    /// Collapses multiple day-of-week buckets for the same hour into a single
-    /// hour-only bucket using pooled statistics.
-    /// </summary>
-    private static BaselineBucket CollapseToHourOnly(List<BaselineBucket> hourBuckets)
-    {
-        var totalSamples = hourBuckets.Sum(b => b.SampleCount);
-        if (totalSamples == 0)
-            return BaselineBucket.Empty;
-
-        // Weighted mean across all day-of-week buckets for this hour
-        var weightedMean = hourBuckets.Sum(b => b.Mean * b.SampleCount) / totalSamples;
-
-        // Pooled standard deviation
-        var pooledVariance = PoolVariance(hourBuckets, weightedMean);
-
-        return new BaselineBucket
-        {
-            HourOfDay = hourBuckets[0].HourOfDay,
-            DayOfWeek = -1, // Indicates hour-only
-            Mean = weightedMean,
-            StdDev = Math.Sqrt(pooledVariance),
-            SampleCount = totalSamples,
-            // Each calendar day lands in exactly one day-of-week bucket, so summing distinct-days
-            // across the dow buckets for this hour is exact (no double-count).
-            DistinctDays = hourBuckets.Sum(b => b.DistinctDays),
-            AbsStdDevFloor = hourBuckets[0].AbsStdDevFloor,
-            Tier = BaselineTier.HourOnly
-        };
-    }
-
-    /// <summary>
-    /// Collapses all buckets into a single flat baseline (equivalent to old 24h behavior).
-    /// </summary>
-    private static BaselineBucket CollapseToFlat(List<BaselineBucket> allBuckets)
-    {
-        var totalSamples = allBuckets.Sum(b => b.SampleCount);
-        if (totalSamples == 0)
-            return BaselineBucket.Empty;
-
-        var weightedMean = allBuckets.Sum(b => b.Mean * b.SampleCount) / totalSamples;
-        var pooledVariance = PoolVariance(allBuckets, weightedMean);
-
-        return new BaselineBucket
-        {
-            HourOfDay = -1,
-            DayOfWeek = -1,
-            Mean = weightedMean,
-            StdDev = Math.Sqrt(pooledVariance),
-            SampleCount = totalSamples,
-            // A calendar day recurs across the 24 hour buckets, so summing would double-count;
-            // MAX is a ~5 ceiling (each (hour, dow) bucket holds at most ~5 same-weekday dates in a
-            // 30-day window) — a cheap proxy that avoids an extra global DISTINCT-days query.
-            DistinctDays = allBuckets.Max(b => b.DistinctDays),
-            AbsStdDevFloor = allBuckets[0].AbsStdDevFloor,
-            Tier = BaselineTier.Flat
-        };
-    }
-
-    /// <summary>
-    /// Computes pooled variance from multiple buckets, accounting for both
-    /// within-bucket variance and between-bucket mean differences.
-    /// </summary>
-    private static double PoolVariance(List<BaselineBucket> buckets, double grandMean)
-    {
-        var totalSamples = buckets.Sum(b => b.SampleCount);
-        if (totalSamples <= 1) return 0;
-
-        double totalSumSq = 0;
-        foreach (var b in buckets)
-        {
-            if (b.SampleCount <= 0) continue;
-            // Within-bucket variance contribution
-            totalSumSq += (b.StdDev * b.StdDev) * (b.SampleCount - 1);
-            // Between-bucket mean difference contribution
-            totalSumSq += b.SampleCount * (b.Mean - grandMean) * (b.Mean - grandMean);
-        }
-
-        return totalSumSq / (totalSamples - 1);
-    }
-
     private class CachedBaseline
     {
         public DateTime ComputedAt { get; init; }
         public DateTime RealTime { get; init; }
         public Dictionary<(int HourOfDay, int DayOfWeek), BaselineBucket>? Buckets { get; init; }
     }
-}
-
-/// <summary>
-/// Represents the computed baseline statistics for a single time bucket.
-/// </summary>
-public class BaselineBucket
-{
-    public int HourOfDay { get; init; }
-    public int DayOfWeek { get; init; }
-    public double Mean { get; init; }
-    public double StdDev { get; init; }
-    public long SampleCount { get; init; }
-    public BaselineTier Tier { get; init; }
-
-    /// <summary>
-    /// Distinct calendar days observed in this bucket — the baseline-QUALITY signal the old
-    /// quantity-only warmup gate lacked (a bucket with many samples but few distinct days is one
-    /// busy day, not a trend). Carried through collapse by CollapseToHourOnly (SUM — each calendar
-    /// day lands in exactly one day-of-week bucket, so the sum is exact) and CollapseToFlat (MAX —
-    /// a calendar day recurs across the 24 hour buckets; MAX is a ~5 ceiling, not the true pooled
-    /// distinct-day count, but a cheap proxy that avoids a second global query).
-    /// </summary>
-    public long DistinctDays { get; init; }
-
-    /// <summary>
-    /// Absolute dispersion floor for BOUNDED metrics (CPU %, memory %, I/O ms) so a
-    /// variance-collapsed baseline can't manufacture a giant z-score. 0 for server-relative
-    /// metrics (batch/query-duration/sessions/waits/blocking), which have no universal floor and
-    /// rely on the detector magnitude floors + the quality gate instead. Set per metric by the provider.
-    /// </summary>
-    public double AbsStdDevFloor { get; init; }
-
-    // Baseline-quality tier gates (see IsTrustworthy). Day-mins are tier-aware: a Full (hour × dow)
-    // bucket only sees ~5 same-weekday dates in a 30-day window, so its day-min is a modest 3. The Flat
-    // tier's DistinctDays is a MAX-over-hour-buckets proxy (CollapseToFlat) capped at that SAME ~5
-    // ceiling, so it can't demand more than a Full bucket — a >=15 floor was structurally unreachable and
-    // left the Flat trust branch permanently dead, so match Full at 3. Sample-mins mirror the provider's
-    // per-tier selection floors.
-    private const long FullSampleMin = 10;
-    private const long FullDayMin = 3;
-    private const long HourOnlySampleMin = 10;
-    private const long HourOnlyDayMin = 10;
-    private const long FlatSampleMin = 3;
-    private const long FlatDayMin = 3;
-
-    public static BaselineBucket Empty => new()
-    {
-        HourOfDay = -1, DayOfWeek = -1, Mean = 0, StdDev = 0,
-        SampleCount = 0, DistinctDays = 0, Tier = BaselineTier.Flat
-    };
-
-    /// <summary>
-    /// Returns the effective stddev with a proportional minimum floor plus, for bounded metrics,
-    /// an absolute floor — both prevent division-by-zero AND a variance-collapsed baseline from
-    /// producing a giant z-score. When both mean and stddev are 0 (zero activity), returns 0 —
-    /// callers should skip scoring (or fall back to the absolute-threshold path).
-    /// </summary>
-    public double EffectiveStdDev
-    {
-        get
-        {
-            if (Mean == 0 && StdDev <= 0) return 0; // Zero activity — skip scoring
-            return Math.Max(Math.Max(StdDev, Mean * 0.01), AbsStdDevFloor);
-        }
-    }
-
-    /// <summary>
-    /// Whether this baseline is dense enough to trust a z-score / ratio against. Requires real
-    /// dispersion, the tier's sample floor, AND enough DISTINCT days. A low-quality baseline is NOT
-    /// silenced — the detector falls back to an absolute-threshold bar instead. This gate and the
-    /// #1486 magnitude floors are COMPLEMENTARY, not both-mandatory: a trustworthy baseline trusts z
-    /// with the magnitude floor as a sanity ceiling; an untrustworthy one fires only on the higher
-    /// absolute bar — they must never AND into blindness on a young store.
-    /// </summary>
-    public bool IsTrustworthy
-    {
-        get
-        {
-            if (EffectiveStdDev <= 0) return false;
-            var (sampleMin, dayMin) = Tier switch
-            {
-                BaselineTier.Full => (FullSampleMin, FullDayMin),
-                BaselineTier.HourOnly => (HourOnlySampleMin, HourOnlyDayMin),
-                _ => (FlatSampleMin, FlatDayMin),
-            };
-            return SampleCount >= sampleMin && DistinctDays >= dayMin;
-        }
-    }
-}
-
-public enum BaselineTier
-{
-    Full,     // hour + day-of-week (168 buckets)
-    HourOnly, // hour only (24 buckets)
-    Flat      // global mean/stddev
-}
-
-/// <summary>Metric name constants used as baseline cache keys.</summary>
-public static class MetricNames
-{
-    public const string Cpu = "cpu";
-    public const string BatchRequests = "batch_requests";
-    public const string WaitStats = "wait_stats";
-    public const string SessionCount = "session_count";
-    public const string QueryDuration = "query_duration";
-    public const string IoLatency = "io_latency";
-    public const string Blocking = "blocking";
-    public const string Deadlock = "deadlock";
-    public const string Memory = "memory";
-
-    // Chart-unit metrics (for UI bands — units match what the chart displays)
-    public const string WaitMsPerSec = "wait_ms_per_sec";
-    public const string BlockingPerMinute = "blocking_per_minute";
 }
