+++ PerformanceMonitor.Analysis/Baselines/AnomalyThresholds.cs
@@ -1,0 +1,74 @@
+/*
+ * Copyright (c) 2026 Erik Darling, Darling Data LLC
+ *
+ * This file is part of the SQL Server Performance Monitor.
+ *
+ * Licensed under the MIT License. See LICENSE file in the project root for full license information.
+ */
+
+namespace PerformanceMonitor.Analysis.Baselines;
+
+/// <summary>
+/// The anomaly-detector tuning constants shared by the two active detectors (Lite
+/// <c>AnomalyDetector</c>, Darling <c>PgAnomalyDetector</c>) so they cannot drift. The detector
+/// METHOD BODIES stay per-store (DuckDB / Npgsql SQL + binding); every numeric threshold, floor,
+/// fallback bar, and sentinel is single-sourced here. The deprecated Dashboard keeps its own copy.
+/// </summary>
+public static class AnomalyThresholds
+{
+    /// <summary>
+    /// Default number of standard deviations above baseline mean to flag as anomalous.
+    /// </summary>
+    public const double DefaultDeviationThreshold = 2.0;
+
+    /// <summary>
+    /// Default ratio threshold for the wait-profile detector (peak window all-types ms/sec ÷ baseline
+    /// mean). On the HONEST per-second scale now, so far below the old 5.0 that assumed a ~240x-inflated
+    /// input; matches the FactScorer WaitProfileRatioFloor. CALIBRATE ON THE SQL2025/HAMMERDB BOX.
+    /// </summary>
+    public const double DefaultRatioThreshold = 4.0;
+
+    /// <summary>
+    /// Default ratio threshold for event-based anomaly detection (blocking/deadlocks).
+    /// </summary>
+    public const double DefaultEventRatioThreshold = 3.0;
+
+    // #1486 absolute-magnitude floors (the z-path sanity ceiling) so a z-score against a thin
+    // baseline can't surface a trivial value; sigma display cap so a variance-collapsed baseline
+    // can't render millions-of-sigma.
+    public const double CpuFloorPct = 50.0;                // %
+    public const double ReadLatencyFloorMs = 10.0;         // ms
+    public const double BatchRequestFloor = 500.0;         // requests/sec
+    public const double SessionCountFloor = 50.0;          // connections
+    public const double QueryDurationFloorUs = 1_000_000;  // total elapsed us = 1 second
+    public const double MemoryPressureFloorPct = 90.0;     // total/target %
+    public const double WriteLatencyFloorMs = 20.0;        // ms, was 5
+    public const double SigmaDisplayCap = 25.0;
+
+    // Low-quality-baseline ABSOLUTE-FALLBACK bars: when the baseline is too thin to trust a z-score
+    // (BaselineBucket.IsTrustworthy false), the detector fires on these instead of going silent.
+    // Each is deliberately HIGHER than the matching #1486 magnitude floor above (the interaction
+    // trap: a young store fires only on the higher bar, never on both-AND-ed into blindness).
+    public const double CpuFallbackPct = 90.0;                 // %
+    public const double MemoryPressureFallbackPct = 95.0;      // total/target %
+    public const double BatchRequestFallback = 5000.0;        // requests/sec
+    public const double SessionCountFallback = 500.0;         // connections
+    public const double QueryDurationFallbackUs = 5_000_000;  // total elapsed us = 5 seconds
+    public const double IoLatencyFallbackMs = 50.0;           // ms (read and write)
+
+    // Wait-profile detector (DetectWaitAnomalies → one ANOMALY_WAIT_PROFILE): the current window's
+    // all-types wait ms/sec (PEAK across collections, matching the z-detectors) is compared to the
+    // WaitMsPerSec baseline. DefaultRatioThreshold and the FactScorer wait slope are on the HONEST
+    // per-second scale now (the old 5×/20× was calibrated to a ~240×-inflated per-hour-vs-per-interval
+    // input) — a sensible starting point; CALIBRATE ON THE SQL2025/HAMMERDB BOX.
+    public const double WaitProfileFallbackMsPerSec = 250.0;  // untrustworthy-baseline absolute bar
+    public const double NoBaselineRatio = 100.0;             // scoring sentinel for a first-occurrence (is_new)
+
+    // Day-over-day object/index detection (delta-based, not stddev-baseline) since the
+    // index_object_stats collector runs daily and its counters are cumulative. Emits
+    // ANOMALY_OBJECT_GROWTH for the biggest table grower over threshold and ANOMALY_OBJECT_CONTENTION
+    // for the index with the largest new lock-wait time.
+    public const decimal ObjectGrowthMbThreshold = 100m;   // ignore tables that grew less than 100 MB
+    public const double ObjectGrowthPctThreshold = 20.0;   // ...and less than 20% day-over-day
+    public const long ObjectLockWaitMsDeltaThreshold = 60000; // 1 minute of new lock waits
+}
