+++ PerformanceMonitor.Analysis/Baselines/MetricNames.cs
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2026 Erik Darling, Darling Data LLC
+ *
+ * This file is part of the SQL Server Performance Monitor.
+ *
+ * Licensed under the MIT License. See LICENSE file in the project root for full license information.
+ */
+
+namespace PerformanceMonitor.Analysis.Baselines;
+
+/// <summary>Metric name constants used as baseline cache keys. Shared by Lite + Darling so the two
+/// active apps key their baselines identically (the deprecated Dashboard keeps its own
+/// <c>SqlServerMetricNames</c>).</summary>
+public static class MetricNames
+{
+    public const string Cpu = "cpu";
+    public const string BatchRequests = "batch_requests";
+    public const string WaitStats = "wait_stats";
+    public const string SessionCount = "session_count";
+    public const string QueryDuration = "query_duration";
+    public const string IoLatency = "io_latency";
+    public const string Blocking = "blocking";
+    public const string Deadlock = "deadlock";
+    public const string Memory = "memory";
+
+    // Chart-unit metrics (for UI bands — units match what the chart displays)
+    public const string WaitMsPerSec = "wait_ms_per_sec";
+    public const string BlockingPerMinute = "blocking_per_minute";
+}
